Explicit casts to avoid warnings

+++ src/Setup/Setup.cpp
@@ -37,13 +37,13 @@ int WINAPI wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance, 
         // and squirrel overheads (incl temp files). the constant 0.38 is a
         // aggressive estimate on what the compression ratio might be.
         int64_t squirrelOverhead = 50 * 1000 * 1000;
-        int64_t requiredSpace = squirrelOverhead + (packageLength * 3) + (packageLength / (double)0.38);
+        int64_t requiredSpace = squirrelOverhead + (packageLength * 3) + (int64_t)((double)packageLength / (double)0.38);
         if (!util::check_diskspace(requiredSpace)) {
             throw wstring(L"Insufficient disk space. This application requires at least " + util::pretty_bytes(requiredSpace) + L" free space to be installed.");
         }
 
         // extract Update.exe and embedded nuget package
-        util::extractUpdateExe(pkgStart, packageLength, updaterPath);
+        util::extractUpdateExe(pkgStart, (size_t)packageLength, updaterPath);
         std::ofstream(packagePath, std::ios::binary).write((char*)pkgStart, packageLength);
 
         // run installer and forward our command line arguments
